LanguageCategoryDisplay: cache last items instead of finding them repeatedly
addToRegion() and getColumn() would call jQuery's find() up to three
times per execution - and they're being executed a few hundred times
in a tight loop.

Since the finds were used simply to find the last inserted elements
we can just cache them in this's properties.

In my profiling on Opera on a Thinkpad T60 (with lazy-loading
disabled), JavaScript evaluation during the rendering of the dialog is
down from ~880ms to ~450 ms.