travis: Switch to Node.js 0.10
This repository is mainly about client-side javascript, the version
of Node.js used for the build scripts doesn't really matter.

One of the dependencies in package.json uses the '^x.x' syntax in
its sub-dependencies and isn't supporteed on Node.js 0.8, which
is why the Travis CI build is currently erroring on the install step.