Adjust styles and code to support custom list bullets
Tweak specificty of the rule that hides bullets so that it is easy
to override. Display bullets inside the item because otherwise they
get hidden because we set overflow: hidden.

Move dir and lang attributes from <li> to <a> so that the bullets
appear on the correct side for reverse directionality scripts.

Change-Id: Ic0d0c5220b0b374c2a3a19023c40d81f3e20007c