Add maxlength to the languagefilter input
This input field is only for language names. There's no reason for it
to have something longer. Looking at the logs of searches,
people sometimes write very long things there: gibberish,
URLs, inappropriate search strings, mistaken pastes from other places,
and so on.

So it should be limited by maxlength.

It's practically never useful after 20 characters, but setting at 40
just to be on the safe side. The longest language name in langdb is 34.